+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Token operations for updating metadata, assigning and revoking admin roles have been added.
+- Token events for updating metadata, assigning and revoking admin roles have been added.
 
 ## 12.0.2